Split CommentFrame from CommentsNode
